Add files via upload
The Antikythera Survey Project (ASP) was a phased, interdisciplinary program of fieldwork, artifact study and laboratory analysis that addressed the long-term history and human ecology of the tiny Greek island of Antikythera. Its program of fieldwork lasted from 2005 to 2008, its artifact study seasons were complete by 2010, and a full set of publications is now available. Please see below and on the other pages of this site for further details.

Antikythera is one of the smallest (ca. 20 sq.km) and most remote inhabited places in the insular Mediterranean, but also one of the best-placed, lying on a key axes of maritime movement, both north-south between the southern Balkan peninsula (the Peloponnese) and Crete, and east-west, between the eastern and central Mediterranean. This strategic, if often fragile and marginal, location is emphasized by the presence of a fortified pirate community on Antikythera during the Hellenistic period (4th to 1st century BC), and a 1st century BC shipwreck a few hundred meters off the coast that famously produced a series of bronze statues and the Antikythera mechanism, an intricately-geared device for maritime navigation.

Antikythera is also an island that provides an incredibly attractive research context, for three main reasons:

small islands are prone to experiencing abrupt demographic changes, including periods of near complete abandonment and recolonisation. Our results have indeed revealed that Antikythera has a long and turbulent history, which stretches back to the latter stages of the Neolithic and includes a substantial Bronze Age presence, a fortified Hellenistic pirate town, several Late Roman communities, Byzantine and Venetian evidence as well as a period of more recent re-colonization. Between several demographic highs are also some apparent demographic crashes, most obviously in the last hundred years. This comparatively discontinuous record of human activity makes it easier to date and understand settlement strategies and multiple cycles of landscape investment (such as terracing) than in most other Mediterranean landscapes.
small islands such as Antikythera play eccentric but extremely revealing roles in wider social, economic and political networks (for example as special places for refugees, hunters, political exiles, hermits, monks and/or pirates). ASP addressed many of these phenomena by combining different types of environmental, archaeological and historical evidence, but also sought to place them in the broader Mediterranean context in which they emerge.
Antikythera is small enough that we were able to cover the entire island by intensive survey and to collect other datasets comprehensively. Whereas high resolution archaeological and environmental survey has been practiced successfully on many other Aegean islands such as Kythera, Melos, Kea and Crete, the projects involved necessarily performed their data collection in a series of restricted sample areas. In contrast, on Antikythera, we were able to collect information across the entire island and this offered a wide variety of analytical opportunities.
These webpages are meant as an introduction to ASP’s aims, methods and results. ASP publications include a book, an open digital data archive and more than ten paper-length discussions which we refer to on other pages of this site where they are relevant. The physical finds from our survey are now stored on the island of Kythera under the direct control of the Greek Archaeological Service and any requests for study access should be made to them. For our part, we welcome further re-study of this material and do not ask to be consulted first, but please do feel free to contact us if you have questions or comments.

Andrew Bevan (UCL)
James Conolly (Trent)